Teach drc-sim-c to decode packets
Pretty impressive numbers: I don't think I wrote this terribly well but it uses only 2% CPU when run with LOG_LEVEL=info. Goes to about 1.7% when in release mode.

There's a lot more software to write around this, but it would have been a crying shame if I couldn't even receive the video packets in a reasonable CPU time!