More changes...
Formatting the WiiUVideoPackets ourselves allows a much more coherent output when verbosity is high.

Logging errors instead of throwing asserts makes this a little easier to use.

Borrowing a slice from the WUP video packet instead of copying the bits is both more efficient and simpler to code.

Changed all of the hardcoded video packet size attributes to a single constant. We might need to tune this later.

Handle more errors in parse_video_packets, like read errors (the buffered reader was stopping reads early when its buffer size was smaller than the read size, or when the two were misaligned)

No need to check the size of recv_from, it is always the size of the buffer.