Garbage collect stale FrameAccumulators
This isn't a perfect implementation, but it is a working one... If we miss any packets in the last ~three frames before the timestamp overflows, we'll leave them for dead until we hit the last ~three frames before the next overflow. Practical downside of this is the simulator might crash every 1.2 hours if we implement it poorly enough.